Move user bio setting to users action/reducer
