+++ shared/actions/chat2-gen.tsx
@@ -122,7 +122,6 @@ export const setThreadLoadStatus = 'chat2:setThreadLoadStatus'
 export const setThreadSearchQuery = 'chat2:setThreadSearchQuery'
 export const setThreadSearchStatus = 'chat2:setThreadSearchStatus'
 export const setUnsentText = 'chat2:setUnsentText'
-export const setUserBio = 'chat2:setUserBio'
 export const setWalletsOld = 'chat2:setWalletsOld'
 export const staticConfigLoaded = 'chat2:staticConfigLoaded'
 export const tabSelected = 'chat2:tabSelected'
@@ -569,7 +568,6 @@ type _SetUnsentTextPayload = {
   readonly conversationIDKey: Types.ConversationIDKey
   readonly text?: HiddenString
 }
-type _SetUserBioPayload = {readonly userCard: RPCTypes.UserCard; readonly username: string}
 type _SetWalletsOldPayload = void
 type _StaticConfigLoadedPayload = {readonly staticConfig: Types.StaticConfig}
 type _TabSelectedPayload = void
@@ -991,13 +989,6 @@ export const createSetUnsentText = (payload: _SetUnsentTextPayload): SetUnsentTe
 export const createSetConvRetentionPolicy = (
   payload: _SetConvRetentionPolicyPayload
 ): SetConvRetentionPolicyPayload => ({payload, type: setConvRetentionPolicy})
-/**
- * Sets user bio for use in one-on-one conversations
- */
-export const createSetUserBio = (payload: _SetUserBioPayload): SetUserBioPayload => ({
-  payload,
-  type: setUserBio,
-})
 /**
  * Static configuration info was loaded from the service.
  */
@@ -1806,7 +1797,6 @@ export type SetUnsentTextPayload = {
   readonly payload: _SetUnsentTextPayload
   readonly type: typeof setUnsentText
 }
-export type SetUserBioPayload = {readonly payload: _SetUserBioPayload; readonly type: typeof setUserBio}
 export type SetWalletsOldPayload = {
   readonly payload: _SetWalletsOldPayload
   readonly type: typeof setWalletsOld
@@ -2037,7 +2027,6 @@ export type Actions =
   | SetThreadSearchQueryPayload
   | SetThreadSearchStatusPayload
   | SetUnsentTextPayload
-  | SetUserBioPayload
   | SetWalletsOldPayload
   | StaticConfigLoadedPayload
   | TabSelectedPayload
