+++ shared/actions/json/users.json
@@ -10,6 +10,11 @@
     },
     "updateFullnames": {
       "usernameToFullname": "{[username: string]: string}"
+    },
+    "updateBio": {
+      "_description": "Sets user bio for use in one-on-one conversations",
+      "userCard": "RPCTypes.UserCard",
+      "username": "string"
     }
   }
 }
