+++ shared/reducers/chat2.tsx
@@ -410,8 +410,6 @@ export default (_state: Types.State = initialState, action: Actions): Types.Stat
       case Chat2Gen.setChannelSearchText:
         draftState.channelSearchText = action.payload.text.toLowerCase()
         return
-      case Chat2Gen.setUserBio:
-        return
       case Chat2Gen.selectConversation:
         {
           // ignore non-changing
