+++ shared/reducers/users.tsx
@@ -4,7 +4,6 @@ import * as Tracker2Gen from '../actions/tracker2-gen'
 import * as Constants from '../constants/users'
 import * as Types from '../constants/types/users'
 import * as ConfigGen from '../actions/config-gen'
-import * as Chat2Gen from '../actions/chat2-gen'
 
 const initialState: Types.State = Constants.makeState()
 const blankUserInfo = Constants.makeUserInfo()
@@ -15,7 +14,6 @@ type Actions =
   | Tracker2Gen.UpdatedDetailsPayload
   | ConfigGen.SetAccountsPayload
   | TeamBuildingGen.SearchResultsLoadedPayload
-  | Chat2Gen.SetUserBioPayload
 
 const reducer = (state: Types.State = initialState, action: Actions): Types.State => {
   switch (action.type) {
@@ -45,6 +43,11 @@ const reducer = (state: Types.State = initialState, action: Actions): Types.Stat
             m.update(user, info => (info || blankUserInfo).set('broken', true))
           })
         })
+      )
+    }
+    case UsersGen.updateBio: {
+      return state.updateIn(['infoMap', action.payload.username], (userInfo = blankUserInfo) =>
+        userInfo.set('bio', action.payload.userCard.bio)
       )
     }
     case Tracker2Gen.updatedDetails:
@@ -83,11 +86,7 @@ const reducer = (state: Types.State = initialState, action: Actions): Types.Stat
           })
         })
       )
-    case Chat2Gen.setUserBio: {
-      return state.updateIn(['infoMap', action.payload.username], (userInfo = blankUserInfo) =>
-        userInfo.set('bio', action.payload.userCard.bio)
-      )
-    }
+
     // Saga only actions
     default:
       return state
